Refactor SharedViewModelFactory
